Fix DocAI Warehouse tests to work properly with a test project (#10526)
[upstream:2b3d1569bfcb8b34144a94509cf77caa8f83d640]
