--- google/services/documentaiwarehouse/resource_document_ai_warehouse_location_generated_test.go
@@ -1,60 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-// ----------------------------------------------------------------------------
-//
-//     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
-//
-// ----------------------------------------------------------------------------
-//
-//     This file is automatically generated by Magic Modules and manual
-//     changes will be clobbered when the file is regenerated.
-//
-//     Please read more about how to change this file in
-//     .github/CONTRIBUTING.md.
-//
-// ----------------------------------------------------------------------------
-
-package documentaiwarehouse_test
-
-import (
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-
-	"github.com/hashicorp/terraform-provider-google/google/acctest"
-)
-
-func TestAccDocumentAIWarehouseLocation_documentAiWarehouseLocationExample(t *testing.T) {
-	t.Parallel()
-
-	context := map[string]interface{}{
-		"random_suffix": acctest.RandString(t, 10),
-	}
-
-	acctest.VcrTest(t, resource.TestCase{
-		PreCheck:                 func() { acctest.AccTestPreCheck(t) },
-		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories(t),
-		Steps: []resource.TestStep{
-			{
-				Config: testAccDocumentAIWarehouseLocation_documentAiWarehouseLocationExample(context),
-			},
-		},
-	})
-}
-
-func testAccDocumentAIWarehouseLocation_documentAiWarehouseLocationExample(context map[string]interface{}) string {
-	return acctest.Nprintf(`
-resource "google_document_ai_warehouse_location" "example" {
-    location = "us"
-    project_number = data.google_project.project.number
-    access_control_mode = "ACL_MODE_DOCUMENT_LEVEL_ACCESS_CONTROL_GCI"
-    database_type = "DB_INFRA_SPANNER"
-    kms_key = "dummy_key"
-    document_creator_default_role = "DOCUMENT_ADMIN"
-}
-
-data "google_project" "project" {
-}
-`, context)
-}
